Handle GitHub attestations.

+++ citester/index.js
@@ -40519,7 +40519,7 @@ async function run() {
                 expectedTags.add(line);
             }
             const tags = new Set();
-            for (const tag of githubPackageRepo.getTags()) {
+            for (const tag of githubPackageRepo.getTags(true)) {
                 tags.add(tag);
             }
             for (const tag of expectedTags) {
@@ -41168,7 +41168,7 @@ __nccwpck_require__.d(__webpack_exports__, {
   "l": () => (/* binding */ GithubPackageRepo)
 });
 
-// UNUSED EXPORTS: ManifestNotFoundException
+// UNUSED EXPORTS: scanRoots
 
 // NAMESPACE OBJECT: ./node_modules/axios/lib/platform/common/utils.js
 var common_utils_namespaceObject = {};
@@ -46541,6 +46541,13 @@ function isValidChallenge(attributes) {
 var jtd = __nccwpck_require__(8876);
 var jtd_default = /*#__PURE__*/__nccwpck_require__.n(jtd);
 ;// CONCATENATED MODULE: ./src/schemas.ts
+const manifestReferenceSchema = {
+    properties: {
+        mediaType: { type: 'string' },
+        digest: { type: 'string' }
+    },
+    additionalProperties: true
+};
 const manifestSchema = {
     properties: {
         mediaType: {
@@ -46550,6 +46557,14 @@ const manifestSchema = {
                 'application/vnd.docker.distribution.manifest.list.v2+json',
                 'application/vnd.docker.distribution.manifest.v2+json'
             ]
+        }
+    },
+    optionalProperties: {
+        manifests: {
+            elements: manifestReferenceSchema
+        },
+        layers: {
+            elements: manifestReferenceSchema
         }
     },
     additionalProperties: true
@@ -46581,43 +46596,34 @@ const packageVersionSchema = {
 };
 
 ;// CONCATENATED MODULE: ./src/models.ts
-var ManifestType;
-(function (ManifestType) {
-    ManifestType[ManifestType["SingleArchitecture"] = 0] = "SingleArchitecture";
-    ManifestType[ManifestType["MultiArchitecture"] = 1] = "MultiArchitecture";
-})(ManifestType || (ManifestType = {}));
-class BaseSingleArchitectureManifest {
-    get type() {
-        return ManifestType.SingleArchitecture;
+class BaseManifest {
+    children;
+    constructor(children = []) {
+        this.children = children;
     }
 }
-class BaseMultiArchitectureManifest {
-    get type() {
-        return ManifestType.MultiArchitecture;
-    }
-}
-class OCIImageManifestModel extends BaseSingleArchitectureManifest {
+class OCIImageManifestModel extends BaseManifest {
     mediaType = 'application/vnd.oci.image.manifest.v1+json';
     constructor(data) {
         super();
         Object.assign(this, data);
     }
 }
-class OCIImageIndexModel extends BaseMultiArchitectureManifest {
+class OCIImageIndexModel extends BaseManifest {
     mediaType = 'application/vnd.oci.image.index.v1+json';
     constructor(data) {
-        super();
+        super(data.manifests ? data.manifests.map(manifest => manifest.digest) : []);
         Object.assign(this, data);
     }
 }
-class DockerManifestListModel extends BaseMultiArchitectureManifest {
+class DockerManifestListModel extends BaseManifest {
     mediaType = 'application/vnd.docker.distribution.manifest.list.v2+json';
     constructor(data) {
-        super();
+        super(data.manifests ? data.manifests.map(manifest => manifest.digest) : []);
         Object.assign(this, data);
     }
 }
-class DockerImageManifestModel extends BaseSingleArchitectureManifest {
+class DockerImageManifestModel extends BaseManifest {
     mediaType = 'application/vnd.docker.distribution.manifest.v2+json';
     constructor(data) {
         super();
@@ -46647,8 +46653,18 @@ class PackageVersionModel {
             tags: []
         }
     });
+    children = [];
+    parent = null;
+    type = 'unknown';
+    manifest = undefined;
     constructor(data) {
         Object.assign(this, data);
+    }
+    get is_attestation() {
+        return this.type === 'attestation root' || this.type === 'attestation child';
+    }
+    toString() {
+        return `{type=${this.type}, id=${this.id}, tags=${this.metadata.container.tags}}`;
     }
 }
 
@@ -46696,11 +46712,102 @@ function parsePackageVersion(jsonString) {
 
 
 
-class ManifestNotFoundException extends Error {
-    constructor(message) {
-        super(message);
-        this.name = 'ManifestNotFoundException';
+function visit(version, fn) {
+    // Visit version.
+    fn(version);
+    // Visit children.
+    for (const child of version.children) {
+        visit(child, fn);
     }
+}
+function scanRoots(uniqueVersions, getVersion) {
+    // Holds versions known to not be roots.
+    const nonRoots = new Set();
+    for (const v of uniqueVersions) {
+        v.children = [];
+        v.parent = null;
+        v.type = 'unknown';
+    }
+    // Loop over all versions.
+    for (const v of uniqueVersions) {
+        // Manifest should have been initialized.
+        if (v.manifest) {
+            // Loop over all digests of child manifests.
+            for (const child_digest of v.manifest.children) {
+                // Get the version for the child digest.
+                const child = getVersion(child_digest);
+                // Check if the child actually exists.
+                if (child && uniqueVersions.has(child)) {
+                    // Add child to v's children.
+                    v.children.push(child);
+                    // The child is not a root version, as per definition.
+                    nonRoots.add(child);
+                    // Add backward reference from child to parent.
+                    child.parent = v;
+                }
+            }
+        }
+    }
+    // Determine roots.
+    const roots = new Set(uniqueVersions);
+    for (const v of nonRoots) {
+        roots.delete(v);
+    }
+    // Set the type for each version.
+    for (const v of uniqueVersions) {
+        if (v.children.length > 0) {
+            v.type = 'multi-arch image';
+            continue;
+        }
+        const m = v.manifest;
+        if (m) {
+            if (m.layers &&
+                m.layers.length === 1 &&
+                m.layers[0].mediaType === 'application/vnd.in-toto+json') {
+                v.type = 'Docker attestation';
+            }
+            else {
+                v.type = 'single-arch image';
+            }
+        }
+        else {
+            v.type = 'unknown';
+        }
+    }
+    // Up to here, we have ignored GitHub attestation versions. They are linked to the version
+    // they attest via a tag that is equivalent to the digest of the attested version. Each attestation
+    // version should be removed from the roots and added to the children of the attested version.
+    // Also, the type should be adjusted for the attestation version and its children.
+    nonRoots.clear();
+    // Loop over all roots.
+    for (const r of roots) {
+        // Loop over all tags for the root.
+        for (const t of r.metadata.container.tags) {
+            // Replace - with : in the tag. If the version is an attestation,
+            // this would be the digest of the attested version.
+            const d = t.replace('-', ':');
+            // Get the version for the potential digest.
+            const v = getVersion(d);
+            // If the version exists and is not identical to r, then r is a GitHub attestation.
+            if (v && v !== r && uniqueVersions.has(v)) {
+                // Recursively set the is_attestation property to true.
+                visit(r, v0 => {
+                    v0.type = 'attestation child';
+                });
+                r.type = 'attestation root';
+                r.parent = v;
+                // Add attestation r to children of v.
+                v.children.push(r);
+                // Add r to non-roots.
+                nonRoots.add(r);
+            }
+        }
+    }
+    // Remove GitHub attestations from root.
+    for (const v of nonRoots) {
+        roots.delete(v);
+    }
+    return roots;
 }
 /**
  * Provides access to a package via the GitHub Packages REST API.
@@ -46710,14 +46817,16 @@ class GithubPackageRepo {
     config;
     // The type of repository (User or Organization)
     repoType = 'Organization';
-    // Maps digest or tag to version.
+    // Maps digest, tag, or id to version.
     versions = new Map();
-    // Maps digest, tag, or id to manifest.
-    manifests = new Map();
     // Collection of tags.
     tags = new Set();
     // Collection of digests.
     digests = new Set();
+    // Collection of unique versions.
+    uniqueVersions = new Set();
+    // Collection of root versions.
+    roots = new Set();
     // HTTP client.
     axios;
     /**
@@ -46802,20 +46911,20 @@ class GithubPackageRepo {
     /**
      * Loads all versions of the package from the GitHub Packages API and populates the internal maps.
      */
-    async mapVersions(fn) {
+    async fetchVersions(fn) {
         // Function to retrieve package versions.
-        let getFunc;
+        let fetch;
         // Parameters for the function call.
-        let getParams;
+        let fetch_params;
         if (this.repoType === 'User') {
             // Use the appropriate function for user repos.
-            getFunc = this.config.isPrivateRepo
+            fetch = this.config.isPrivateRepo
                 ? this.config.octokit.rest.packages
                     .getAllPackageVersionsForPackageOwnedByAuthenticatedUser
                 : this.config.octokit.rest.packages
                     .getAllPackageVersionsForPackageOwnedByUser;
             // Parameters for the function call.
-            getParams = {
+            fetch_params = {
                 package_type: 'container',
                 package_name: this.config.package,
                 username: this.config.owner,
@@ -46824,11 +46933,11 @@ class GithubPackageRepo {
             };
         }
         else {
-            getFunc =
+            fetch =
                 this.config.octokit.rest.packages
                     .getAllPackageVersionsForPackageOwnedByOrg;
             // Parameters for the function call.
-            getParams = {
+            fetch_params = {
                 package_type: 'container',
                 package_name: this.config.package,
                 org: this.config.owner,
@@ -46837,35 +46946,38 @@ class GithubPackageRepo {
             };
         }
         // Iterate over all package versions.
-        for await (const response of this.config.octokit.paginate.iterator(getFunc, getParams)) {
+        for await (const response of this.config.octokit.paginate.iterator(fetch, fetch_params)) {
             for (const packageVersion of response.data) {
                 const version = parsePackageVersion(JSON.stringify(packageVersion));
-                // Get the manifest for the package version.
                 const manifest = await this.fetchManifest(version.name);
-                fn(version, manifest);
+                version.manifest = manifest;
+                fn(version);
             }
         }
     }
-    addVersion(version, manifest) {
+    addVersion(version) {
         this.versions.set(version.name, version);
         this.versions.set(version.id, version);
         this.digests.add(version.name);
-        this.manifests.set(version.name, manifest);
-        this.manifests.set(version.id, manifest);
         // Add each tag to the internal map.
         for (const tag of version.metadata.container.tags) {
             this.versions.set(tag, version);
             this.tags.add(tag);
-            this.manifests.set(tag, manifest);
         }
+        this.uniqueVersions.add(version);
+    }
+    getRoots() {
+        return this.roots;
     }
     async loadVersions() {
         // Clear the internal maps.
         this.versions.clear();
-        this.manifests.clear();
+        this.uniqueVersions.clear();
         this.tags.clear();
         this.digests.clear();
-        return this.mapVersions(this.addVersion.bind(this));
+        await this.fetchVersions(this.addVersion.bind(this));
+        this.roots = scanRoots(this.uniqueVersions, this.getVersion.bind(this));
+        return Promise.resolve();
     }
     /**
      Return the digests for the package.
@@ -46874,12 +46986,22 @@ class GithubPackageRepo {
     getDigests() {
         return Array.from(this.digests);
     }
+    getVersions() {
+        // filter duplicates
+        return Array.from(new Set(this.versions.values()));
+    }
     /**
      * Return the tags for the package.
      * @returns The tags for the package.
      */
-    getTags() {
-        return Array.from(this.tags);
+    getTags(includeAttestations = false) {
+        const tags = Array.from(this.tags);
+        if (includeAttestations) {
+            return tags;
+        }
+        else {
+            return tags.filter(tag => !this.getVersion(tag)?.is_attestation);
+        }
     }
     /**
      * Return the package version for a tag.
@@ -46929,43 +47051,26 @@ class GithubPackageRepo {
         // Remove digest from internal set.
         this.digests.delete(version.name);
         this.versions.delete(version.name);
-        this.manifests.delete(version.name);
         // Remove tags from internal set.
         for (const tag of version.metadata.container.tags) {
             this.tags.delete(tag);
             this.versions.delete(tag);
-            this.manifests.delete(tag);
         }
         this.versions.delete(id);
-        this.manifests.delete(id);
+        this.uniqueVersions.delete(version);
+        this.roots = scanRoots(this.uniqueVersions, this.getVersion.bind(this));
     }
     /**
      * Retrieves a manifest by its digest.
      *
      * @param digest - The digest of the manifest to retrieve.
      * @returns A Promise that resolves to the retrieved manifest.
-     * @throws {ManifestNotFoundException} If the manifest is not found for the given digest.
      */
     async fetchManifest(digest) {
-        try {
-            // Retrieve the manifest.
-            const response = await this.axios.get(`/v2/${this.config.owner}/${this.config.package}/manifests/${digest}`);
-            const manifest = parseManifest(JSON.stringify(response?.data));
-            return manifest;
-        }
-        catch (error) {
-            if (axios_isAxiosError(error) &&
-                error.response != null &&
-                error.response.status === 400) {
-                throw new ManifestNotFoundException(`Manifest not found for digest ${digest}`);
-            }
-            else {
-                throw error;
-            }
-        }
-    }
-    getManifest(key) {
-        return this.manifests.get(key);
+        // Retrieve the manifest.
+        const response = await this.axios.get(`/v2/${this.config.owner}/${this.config.package}/manifests/${digest}`);
+        const manifest = parseManifest(JSON.stringify(response?.data));
+        return manifest;
     }
     /**
      * Puts the manifest for a given tag in the registry.
@@ -46986,7 +47091,7 @@ class GithubPackageRepo {
             try {
                 // Try to put the manifest without token.
                 const response = await auth.put(`https://ghcr.io/v2/${this.config.owner}/${this.config.package}/manifests/${tag}`, manifest, config);
-                core.info(`New digest: ${response.headers['docker-content-digest']}`);
+                core.debug(`New digest: ${response.headers['docker-content-digest']}`);
             }
             catch (error) {
                 if (axios_isAxiosError(error) && error.response?.status === 401) {
@@ -46997,7 +47102,7 @@ class GithubPackageRepo {
                             Authorization: `Bearer ${token}`
                         }
                     });
-                    core.info(`New digest: ${response.headers['docker-content-digest']}`);
+                    core.debug(`New digest: ${response.headers['docker-content-digest']}`);
                 }
                 else {
                     throw error;
@@ -47008,48 +47113,49 @@ class GithubPackageRepo {
     async deleteTag(tag) {
         // Get the version for the tag.
         const version = this.getVersion(tag);
-        // Get the manifest for the tag.
-        const manifest = this.getManifest(tag);
-        if (!version || !manifest) {
+        if (!version) {
             throw new Error(`Version or manifest not found for tag ${tag}.`);
         }
-        // Clone the manifest.
-        const manifest0 = JSON.parse(JSON.stringify(manifest));
-        // Make manifest0 into a fake manifest that does not point to any other manifests or layers.
-        // Push the manifest with the given tag to the registry. This creates a new version with the
-        // tag and removes it from the original version.
-        if (manifest0.manifests) {
-            // Multi-arch manifest. Remove any pointers to child manifests.
-            manifest0.manifests = [];
+        if (!this.config.dryRun) {
+            // Clone the manifest.
+            const manifest0 = JSON.parse(JSON.stringify(version.manifest));
+            // Make manifest0 into a fake manifest that does not point to any other manifests or layers.
+            // Push the manifest with the given tag to the registry. This creates a new version with the
+            // tag and removes it from the original version.
+            if (manifest0.manifests) {
+                // Multi-arch manifest. Remove any pointers to child manifests.
+                manifest0.manifests = [];
+            }
+            else {
+                // Single-architecture or attestation manifest. Remove any pointers to layers.
+                manifest0.layers = [];
+            }
+            await this.putManifest(tag, manifest0);
         }
-        else {
-            // Single-architecture or attestation manifest. Remove any pointers to layers.
-            manifest0.layers = [];
-        }
-        await this.putManifest(tag, manifest0);
         // Remove the tag from the original version. The original manifest does not need to be adjusted.
         version.metadata.container.tags = version.metadata.container.tags.filter(t => t !== tag);
         // It should not be necessary, but just to remove the tag cleanly at this point, remove it from the internal maps and sets.
         this.tags.delete(tag);
         this.versions.delete(tag);
-        this.manifests.delete(tag);
-        // Fetch the new version for the tag.
-        const fn = (version_, manifest_) => {
-            if (version_.metadata.container.tags.includes(tag)) {
-                this.addVersion(version_, manifest_);
+        if (!this.config.dryRun) {
+            // Fetch the new version for the tag.
+            const fn = (version_) => {
+                if (version_.metadata.container.tags.includes(tag)) {
+                    this.addVersion(version_);
+                }
+            };
+            // Reload the package repository to update the version cache.
+            await this.fetchVersions(fn.bind(this));
+            // Get the new version for the tag.
+            const version0 = this.getVersion(tag);
+            if (version0) {
+                core.debug(JSON.stringify(version0, null, 2));
+                // Delete the temporary version.
+                await this.deleteVersion(tag);
             }
-        };
-        // Reload the package repository to update the version cache.
-        await this.mapVersions(fn.bind(this));
-        // Get the new version for the tag.
-        const version0 = this.getVersion(tag);
-        if (version0) {
-            core.debug(JSON.stringify(version0, null, 2));
-            // Delete the temporary version.
-            await this.deleteVersion(tag);
-        }
-        else {
-            throw new Error(`Intermediate version used to delete tag ${tag} not found.`);
+            else {
+                throw new Error(`Intermediate version used to delete tag ${tag} not found.`);
+            }
         }
     }
 }
